+++ doc/releaseNotes.md
@@ -388,3 +388,9 @@ This file describes the main feature changes for each readout.exe released versi
 
 ## v2.6.1 - 11/10/2021
 - Fix monitoring statistics: memlocked pages.
+
+## v2.7.0 - 19/10/201
+- Added linkId in RDH error reporting.
+- Error reported when first orbit received by equipments mismatch.
+- Updated configuration parameters:
+  - readout.maxMsgError and readout.maxMsgWarning: when set, readout will stop when reaching the number of error or warning messages defined as threshold.